+++ src/system/mod.rs
@@ -2,3 +2,4 @@
 
 pub(crate) mod environment;
 pub(crate) mod fs;
+pub(crate) mod steam;